[PATCH] fix show_refcnt return value type
module_attribute.show is defined to return ssize_t

+++ kernel/module.c
@@ -649,7 +649,7 @@ void symbol_put_addr(void *addr)
 }
 EXPORT_SYMBOL_GPL(symbol_put_addr);
 
-static int show_refcnt(struct module *mod, char *buffer)
+static ssize_t show_refcnt(struct module *mod, char *buffer)
 {
 	/* sysfs holds a reference */
 	return sprintf(buffer, "%u\n", module_refcount(mod)-1);
